requirements.txt
beautifulsoup4>=4.14
openpyxl>=3.1
pandas>=2.2
pdfplumber>=0.11
python-docx>=1.1
requests>=2.32
